CM-487: move files to components
